+++ src/Servant/RawM/Internal/Client.hs
@@ -25,10 +25,13 @@ import Network.HTTP.Types (Header, Method)
 import Servant.Client (Client, ClientM, HasClient(clientWithRoute))
 import Servant.Common.Req (Req, performRequest)
 
-import Servant.RawM.Internal.API (RawM)
+import Servant.RawM.Internal.API (RawM')
 
-instance HasClient RawM where
-  type Client RawM = Method -> (Req -> Req) -> ClientM (Int, ByteString, MediaType, [Header], Response ByteString)
+instance HasClient (RawM' serverType) where
+  type Client (RawM' serverType) =
+        Method
+    -> (Req -> Req)
+    -> ClientM (Int, ByteString, MediaType, [Header], Response ByteString)
 
-  clientWithRoute :: Proxy RawM -> Req -> Client RawM
+  clientWithRoute :: Proxy (RawM' serverType) -> Req -> Client (RawM' serverType)
   clientWithRoute Proxy req method f = performRequest method $ f req