+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 
+## 0.2.0.0
+
+*   (commit 30a2cd48488d) Add a phantom type to `RawM` to allow the user to
+    change the output documentation.
+*   (commit 30a2cd48488d) Add a `HasDocs` instance for `RawM` that uses
+    the `HasDocs` instance for the phantom type.
+
 ## 0.1.0.0
 
 *   Initial release.